docker/mqtt-prediction/requirements.txt
paho-mqtt
pandas
joblib
